Make MQTT client id configurable
Fix #373 for two adapters conflicting on the broker with the same client
id. Use the MQTTDeviceID instead of a hard coded default. If the option
is unset, use "BSB-LAN" for backwards compatibility (in most cases
anyway).

Attention: Breaks compatibility if and only if you already have a custom
MQTTDeviceID *and* your broker relies on the client ID in any way for
authorization purposes. In almost all other cases, the change will not
be noticeable.