Fixes in mqtt_handler.h
An exception was thrown after publishing a json message in mqtt_mode 3.
Apparently a buffer overlow in outBuf was responsible for the exception,
since it did not happen for mqtt_mode 1 or 2, where the messages are
significantly shorter. The refactored function mqtt_sendtoBroker() is
not using outBuf anymore.

Also in the buffer overflow protection of function mqtt_callback(...) 4
bytes were reserved for creating the acknowledgement message from the
original message, but 5 bytes are needed (4 for the ACK_ prefix and one
for the terminating nul character).