Simplify Changes to Units
This essentially reverts many of the file changes made when creating a separate MQTT unit option, whilst maintaining the same new functionality.

The `len` field of unit_str_len_t was never used anywhere, so there's no point keeping it, and this then allows must of the files to go back to how they were.